+++ src-tauri/src/lib.rs
@@ -11,9 +11,14 @@ pub fn run() {
             
             // 延迟显示窗口，等待内容加载
             std::thread::spawn(move || {
+                #[cfg(not(debug_assertions))]
                 std::thread::sleep(std::time::Duration::from_millis(400));
                 main_window.show().unwrap();
                 main_window.set_focus().unwrap();
+                
+                // 在开发环境自动打开开发者工具
+                #[cfg(debug_assertions)]
+                main_window.open_devtools();
             });
 
             Ok(())