+++ src/main.ts
@@ -3,7 +3,7 @@ import App from "./App.vue";
 import router from "./router";
 // 全局样式：放大基础字体和 Element Plus 基础字号
 import "./assets/styles/global.css";
-import { updateApiConfig } from "./server";
+import { setAdminApiKey, updateApiConfig } from "./server";
 import { fetch as tauriFetch } from '@tauri-apps/plugin-http';
 
 // 仅在 Tauri 环境下使用 plugin-http 与直连后端；
@@ -12,8 +12,10 @@ import { fetch as tauriFetch } from '@tauri-apps/plugin-http';
 updateApiConfig({
     baseURL: 'http://127.0.0.1:8080',
     timeoutMs: 10000,
-    customFetch: tauriFetch
+    customFetch: tauriFetch,
+    isAdminMode: true,
 });
+setAdminApiKey("ADMIN_KEY_3f6e40cb43b742a0894754866c2e1abe");
 
 const app = createApp(App);
 app.use(router);